requirements.txt
pytest==7.2.0
pytest-cov==4.0.0
requests==2.28.1
google-cloud-storage==2.6.0
pysam==0.19.1
